requirements.txt
# ClawForge v4.0 — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework & API ──────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1

# ── System Monitoring ───────────────────────
psutil==5.9.8

# ── Database ────────────────────────────────
sqlalchemy==2.0.30

# ── Office Documents ────────────────────────
python-docx==1.1.2
openpyxl==3.1.3
reportlab==4.2.0

# ── HTTP Client ─────────────────────────────
requests==2.32.3
httpx==0.27.0

# ── Security ────────────────────────────────
cryptography==42.0.7

# ── UI Automation (screen control) ──────────
pyautogui==0.9.54
pillow==10.3.0

# ── Web Scraping (SAFE mode only) ───────────
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Data Processing ─────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── CLI & Formatting ────────────────────────
rich==13.7.1
click==8.1.7

# ── Scheduling (automation tasks) ───────────
schedule==1.2.2

# ── Environment ─────────────────────────────
python-dotenv==1.0.1

# ── Dev / Testing ───────────────────────────
pytest==8.2.1
httpx==0.27.0
